requirements.txt
PyMuPDF
gradio
numpy
sentence-transformers
tiktoken
torch
transformers
accelerate
llama_cpp_python